Ported packet to new rrsets [please-review], temporary.
The rdata-from-wire parsing is very similar to what compiler
does, so the code should be probably merged and reused.
Also there seems to be some problem with parsing NAPTR.

+++ src/libknot/packet/packet.h
@@ -370,7 +370,7 @@ uint16_t knot_packet_qtype(const knot_packet_t *packet);
  * \param packet Packet containing question.
  * \param qtype New QTYPE for question.
  */
-void knot_packet_set_qtype(knot_packet_t *packet, int qtype);
+void knot_packet_set_qtype(knot_packet_t *packet, uint16_t qtype);
 
 
 /*!
